+++ src/dom.rs
@@ -10,6 +10,7 @@ use crate::network::url::Url;
 use crate::platform::Platform;
 use crate::resource_loader::{
     self,
+    CookieStore,
     ResourceRequestJobTracker,
     ResourceRequestResult,
     ResourceThreadPool,
@@ -39,10 +40,10 @@ impl Document {
         return Document { document_node: Rc::from(RefCell::from(ElementDomNode::new_empty())),
             all_nodes: HashMap::new(), style_context: StyleContext { user_agent_sheet: vec![], author_sheet: vec![] }, base_url: Url::empty() };
     }
-    pub fn update_all_dom_nodes(&mut self, resource_thread_pool: &mut ResourceThreadPool) -> bool {
+    pub fn update_all_dom_nodes(&mut self, resource_thread_pool: &mut ResourceThreadPool, cookie_store: &CookieStore) -> bool {
         //returns whether there are dirty nodes after the update
 
-        return self.document_node.borrow_mut().update(resource_thread_pool, self);
+        return self.document_node.borrow_mut().update(resource_thread_pool, self, cookie_store);
     }
     pub fn find_parent_with_name(&self, start_node: &ElementDomNode, name_to_match: &str) -> Option<Rc<RefCell<ElementDomNode>>> {
         let mut node_id_to_check = start_node.parent_id;
@@ -190,14 +191,14 @@ impl ElementDomNode {
             }
         }
     }
-    fn update(&mut self, resource_thread_pool: &mut ResourceThreadPool, document: &Document) -> bool {
+    fn update(&mut self, resource_thread_pool: &mut ResourceThreadPool, document: &Document, cookie_store: &CookieStore) -> bool {
         //returns whether there are dirty nodes after the update (being itself, or any of the children)
 
         let mut any_child_dirty = false;
 
         if self.children.is_some() {
             for child in self.children.as_ref().unwrap() {
-                let child_dirty = child.borrow_mut().update(resource_thread_pool, document);
+                let child_dirty = child.borrow_mut().update(resource_thread_pool, document, cookie_store);
                 if child_dirty {
                     any_child_dirty = true;
                 }
@@ -211,8 +212,8 @@ impl ElementDomNode {
                 if self.img_job_tracker.is_none() {
                     let image_url = Url::from_base_url(&image_src.unwrap(), Some(&document.base_url));
 
-                    self.img_job_tracker = Some(resource_loader::schedule_load_image(&image_url, resource_thread_pool)); //TODO: eventually store the threadpool
-                                                                                                                         //      on a more general context object
+                     //TODO: eventually store the threadpool on a more general context object
+                    self.img_job_tracker = Some(resource_loader::schedule_load_image(&image_url, &cookie_store, resource_thread_pool));
 
                 } else {
                     let try_recv_result = self.img_job_tracker.as_ref().unwrap().receiver.try_recv();