+++ HISTORY.md
@@ -1,6 +1,7 @@
 0.6.0 [UNRELEASED]
 ===================
-- Improved content selection logic
+- Improved content selection logic\
+- Basic cookie support
 
 
 0.5.0